% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_fam_1}
\alias{test_fam_1}
\title{test_fam_1}
\format{
An object of class \code{data.frame} with 19 rows and 23 columns.
}
\usage{
test_fam_1
}
\description{
A complete example of pedigree data frame with 19 relatives and
family history of breast and ovarian cancer. It also includes germline
testing results for BRCA1 and BRCA2, as well as tumor biomarker testing for
ER, CK5.6, CK14, PR, and HER.
}
\keyword{datasets}
