% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppToVisNetwork.R
\name{toVisNetwork}
\alias{toVisNetwork}
\title{Converts a PanelPRO compatible data frame into a visNetwork data frame}
\usage{
toVisNetwork(ped, title = "Your Pedigree")
}
\arguments{
\item{ped}{a pedigree that contains \code{ID, Sex, MotherID, FatherID, isProband, CurAge}
It is assumed that in this data.frame, Sex = 1 if male, Sex = 0 if female and Sex = NA if unkonwn
Affliction is coded in "isAffXX" columns, where XX are short names of cancers
The age of diagnosis is correspondingly coded in "AgeXX" columns}

\item{title}{a string for the title on the plot}
}
\description{
Converts a PanelPRO compatible data frame into a visNetwork data frame
}
\examples{
toVisNetwork(test_fam_1)
}
