% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visPed.R
\name{visEngine}
\alias{visEngine}
\title{Visualization engine inspired by kinship2}
\usage{
visEngine(
  x,
  annot,
  feature.name = NULL,
  which.proband = NULL,
  add.ids = x$id[x$add == 1],
  id = x$id,
  status = x$status,
  affected = x$affected,
  cex = 1,
  col = 1,
  symbolsize = 1,
  branch = 0.6,
  packed = TRUE,
  align = c(1.5, 2),
  width = 10,
  density = c(-1, 35, 65, 20),
  main_title = "Your Pedigree",
  mar = c(4.1, 1, 4.1, 1),
  angle = c(90, 65, 40, 0),
  keep.par = FALSE,
  subregion,
  pconnect = 0.5
)
}
\arguments{
\item{x}{a pedigree data.frame}

\item{annot}{the annotation string}

\item{feature.name}{the names of the features}

\item{which.proband}{which ID is the proband}

\item{add.ids}{which additional IDs to add}

\item{id}{ID column from x pedigree}

\item{status}{alive or dead status}

\item{affected}{affected or unaffected column}

\item{cex}{spacing}

\item{col}{column number, either 1 or the number of individuals}

\item{symbolsize}{symbol size}

\item{branch}{branch size}

\item{packed}{boolean to pack the pedigree plot}

\item{align}{align parameters}

\item{width}{width parameter}

\item{density}{density parameters}

\item{main_title}{title on plot}

\item{mar}{legend margins}

\item{angle}{angle of labels}

\item{keep.par}{boolean to keep or discard old graphics parameters}

\item{subregion}{boolean whether or not to split plot into subregions}

\item{pconnect}{controls parent connections}
}
\description{
Visualization engine inspired by kinship2
}
